// cashfree-node-pg-sdk/model/apiError502.ts
/**
 * Cashfree Payment Gateway APIs
 * Cashfree\'s Payment Gateway APIs provide developers with a streamlined pathway to integrate advanced payment processing capabilities into their applications, platforms and websites.
 *
 * The version of the OpenAPI document: 2022-01-01
 * Contact: developers@cashfree.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Error when there is error at partner bank
*/
export class ApiError502 {
    'message'?: string;
    /**
    * `bank_processing_failure` will be returned here to denote failure at bank. 
    */
    'code'?: string;
    /**
    * api_error
    */
    'type'?: ApiError502.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "ApiError502.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return ApiError502.attributeTypeMap;
    }
}

export namespace ApiError502 {
    export enum TypeEnum {
        ApiError = <any> 'api_error'
    }
}