// cashfree-node-pg-sdk/model/customerDetails.ts
/**
 * Cashfree Payment Gateway APIs
 * Cashfree\'s Payment Gateway APIs provide developers with a streamlined pathway to integrate advanced payment processing capabilities into their applications, platforms and websites.
 *
 * The version of the OpenAPI document: 2022-01-01
 * Contact: developers@cashfree.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* The customer details that are necessary. Note that you can pass dummy details if your use case does not require the customer details.
*/
export class CustomerDetails {
    /**
    * A unique identifier for the customer. Use alphanumeric values only.
    */
    'customerId': string;
    /**
    * Customer email address.
    */
    'customerEmail'?: string | null;
    /**
    * Customer phone number.
    */
    'customerPhone': string;
    /**
    * Name of the customer.
    */
    'customerName'?: string | null;
    /**
    * Customer bank account. Required if you want to do a bank account check (TPV)
    */
    'customerBankAccountNumber'?: string | null;
    /**
    * Customer bank IFSC. Required if you want to do a bank account check (TPV)
    */
    'customerBankIfsc'?: string | null;
    /**
    * Customer bank code. Required for net banking payments, if you want to do a bank account check (TPV)
    */
    'customerBankCode'?: number | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "customerId",
            "baseName": "customer_id",
            "type": "string"
        },
        {
            "name": "customerEmail",
            "baseName": "customer_email",
            "type": "string"
        },
        {
            "name": "customerPhone",
            "baseName": "customer_phone",
            "type": "string"
        },
        {
            "name": "customerName",
            "baseName": "customer_name",
            "type": "string"
        },
        {
            "name": "customerBankAccountNumber",
            "baseName": "customer_bank_account_number",
            "type": "string"
        },
        {
            "name": "customerBankIfsc",
            "baseName": "customer_bank_ifsc",
            "type": "string"
        },
        {
            "name": "customerBankCode",
            "baseName": "customer_bank_code",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CustomerDetails.attributeTypeMap;
    }
}